MaterialVisual/requirements.txt
django-filter==21.1
djangorestframework==3.12.4
Django==3.2.6
pillow==8.4.0
djangorestframework-jwt==1.11.0
coreapi==2.3.3
pymysql==1.0.2
uwsgi==2.0.20
python-ldap==3.3.1
django-auth-ldap==3.0.0
django-simple-history==3.0.0
django-crispy-forms==1.14.0
django-reversion==5.0.2
django-formtools==2.3
django-import-export==2.8.0
django-ckeditor==6.3.2
httplib2==0.20.4
future==0.18.2
six==1.12.0
xlwt==1.3.0
xlsxwriter==3.0.3
PyYAML==6.0
